+++ pkg/config/image_test.go
@@ -1,0 +1,62 @@
+// Copyright 2023 The Okteto Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package config
+
+import (
+	"testing"
+
+	"github.com/okteto/okteto/pkg/log/io"
+	"github.com/stretchr/testify/require"
+)
+
+func Test_getOktetoCLIVersion(t *testing.T) {
+	var tests = []struct {
+		name                                 string
+		versionString, expected, cliImageEnv string
+	}{
+		{
+			name:          "no version string and no env return latest",
+			versionString: "",
+			expected:      "okteto/okteto:latest",
+		},
+		{
+			name:          "no version string return env value",
+			versionString: "",
+			cliImageEnv:   "okteto/remote:test",
+			expected:      "okteto/remote:test",
+		},
+		{
+			name:          "found version string",
+			versionString: "2.2.2",
+			expected:      "okteto/okteto:2.2.2",
+		},
+		{
+			name:          "found incorrect version string return latest ",
+			versionString: "2.a.2",
+			expected:      "okteto/okteto:latest",
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			if tt.cliImageEnv != "" {
+				t.Setenv(oktetoDeployRemoteImageEnvVar, tt.cliImageEnv)
+			}
+
+			version := NewImageConfig(io.NewIOController()).GetRemoteImage(tt.versionString)
+			require.Equal(t, version, tt.expected)
+		})
+	}
+}
