+++ pkg/remote/remote_test.go
@@ -549,6 +549,7 @@ COPY --from=runner /etc/.oktetocachekey .oktetocachekey
 				getEnviron: func() []string {
 					return []string{}
 				},
+				ioCtrl:             io.NewIOController(),
 				generateSocketName: nameGenerator.GenerateName,
 			}
 			dockerfileName, err := rdc.createDockerfile("/test", tt.config.params)
@@ -582,6 +583,7 @@ func TestDockerfileWithCache(t *testing.T) {
 		getEnviron: func() []string {
 			return []string{}
 		},
+		ioCtrl:             io.NewIOController(),
 		generateSocketName: nameGenerator.GenerateName,
 	}
 	caches := []string{"/my", "/cache", "/list"}
@@ -598,47 +600,6 @@ func TestDockerfileWithCache(t *testing.T) {
 		ok, err := regexp.MatchString(pattern, string(d))
 		require.NoError(t, err)
 		require.True(t, ok)
-	}
-}
-
-func Test_getOktetoCLIVersion(t *testing.T) {
-	var tests = []struct {
-		name                                 string
-		versionString, expected, cliImageEnv string
-	}{
-		{
-			name:          "no version string and no env return latest",
-			versionString: "",
-			expected:      "okteto/okteto:latest",
-		},
-		{
-			name:          "no version string return env value",
-			versionString: "",
-			cliImageEnv:   "okteto/remote:test",
-			expected:      "okteto/remote:test",
-		},
-		{
-			name:          "found version string",
-			versionString: "2.2.2",
-			expected:      "okteto/okteto:2.2.2",
-		},
-		{
-			name:          "found incorrect version string return latest ",
-			versionString: "2.a.2",
-			expected:      "okteto/okteto:latest",
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.cliImageEnv != "" {
-				t.Setenv(constants.OktetoDeployRemoteImage, tt.cliImageEnv)
-			}
-
-			version := getOktetoCLIVersion(tt.versionString)
-			require.Equal(t, version, tt.expected)
-		})
 	}
 }
